+++ src/main/java/com/bookshop/repository/BookRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.bookshop.domain.Book;
 
-public interface BookRepository extends PagingAndSortingRepository<Book, Long> {
+public interface BookRepository extends PagingAndSortingRepository<Book, Integer> {
 
 }
